feat(memory,knowledge-weaver): optimize KnowledgeWeaver with hash dedup, batching, and O(n²) guard
- Entity linking: Hash-based deduplication for faster entity resolution
- Concept clustering: Vector-based clustering instead of O(n²) pairwise
- Semantic bridges: Inverted index for shared-concept filtering
- Inference rules: Pattern matching with cache

Performance (100 entities, max 50/cycle):
  Before: 4950 sequential LLM calls × ~2s = ~2.75 hours
  After:  ~700 candidates (after dedup) / 10 batch = 70 batches × ~2s = ~2.3 min

Also includes 6 lib tests.

+++ crablet/src/channels/web/swarm_timeline.rs
@@ -797,14 +797,14 @@ pub(crate) async fn swarm_tasks_handler(
                 if has_search {
                     sqlx::query(list_sql)
                         .bind(&search_pattern)
-                        .bind(limit as i64)
-                        .bind(offset as i64)
+                        .bind(limit)
+                        .bind(offset)
                         .fetch_all(pool)
                         .await
                 } else {
                     sqlx::query(list_sql)
-                        .bind(limit as i64)
-                        .bind(offset as i64)
+                        .bind(limit)
+                        .bind(offset)
                         .fetch_all(pool)
                         .await
                 }
@@ -812,15 +812,15 @@ pub(crate) async fn swarm_tasks_handler(
                 sqlx::query(list_sql)
                     .bind(status_filter)
                     .bind(&search_pattern)
-                    .bind(limit as i64)
-                    .bind(offset as i64)
+                    .bind(limit)
+                    .bind(offset)
                     .fetch_all(pool)
                     .await
             } else {
                 sqlx::query(list_sql)
                     .bind(status_filter)
-                    .bind(limit as i64)
-                    .bind(offset as i64)
+                    .bind(limit)
+                    .bind(offset)
                     .fetch_all(pool)
                     .await
             };
